fix(node): add exponential backoff for gateway connection retries
When gateway connections fail, the node was retrying immediately without
any delay. This caused rapid retry loops that:
1. Generated excessive log spam (filling up disk space)
2. Wasted network resources with futile connection attempts
3. Made debugging difficult due to log noise

This change integrates the existing PeerConnectionBackoff module into
the connection retry path. When a connection fails:
- Record the failure to trigger exponential backoff
- Wait the appropriate delay (5s base, doubling up to 5min max)
- Then retry the connection

The backoff is tracked per-gateway by socket address and automatically
expires after successful periods without failures.

Also fixes unrelated clippy warnings in ledbat.rs tests.

+++ crates/core/src/transport/ledbat.rs
@@ -4708,10 +4708,10 @@ mod tests {
         // Phase 1: Establish original base delay at 20ms
         let old_rtt = Duration::from_millis(20);
         println!("\n--- Phase 1: Original path @ 20ms ---");
-        for i in 0..BASE_HISTORY_SIZE as usize {
+        for i in 0..BASE_HISTORY_SIZE {
             controller.on_ack(old_rtt, 3000);
             tokio::time::sleep(Duration::from_millis(25)).await;
-            if i == BASE_HISTORY_SIZE as usize - 1 {
+            if i == BASE_HISTORY_SIZE - 1 {
                 println!(
                     "  Established base_delay: {:?}, cwnd: {} KB",
                     controller.base_delay(),
@@ -4737,7 +4737,7 @@ mod tests {
 
         // Need enough samples to age out the old base delay
         // The base delay history keeps BASE_HISTORY_SIZE samples
-        for i in 0..(BASE_HISTORY_SIZE as usize * 2) {
+        for i in 0..(BASE_HISTORY_SIZE * 2) {
             tokio::time::sleep(Duration::from_millis(85)).await;
             controller.on_ack(new_rtt, 3000);
